Maps added and bugs fixed

+++ backend/destinationsBackend/src/main/java/com/example/destinationsbackend/Place/PlaceService.java
@@ -2,10 +2,10 @@ package com.example.destinationsbackend.Place;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.CrossOrigin;
 
 import java.util.List;
 import java.util.Optional;
+
 @Service
 public class PlaceService {
 
@@ -15,17 +15,24 @@ public class PlaceService {
     public Optional<Place> getPlaceById(Long id) {
         return placeJpaRepository.findById(id);
     }
-    public List<Place> getAllPlacesByUserId(Long id) {
-        return placeJpaRepository.findAllByUserId(id);
+
+    public List<Place> getAllPlacesByUserId(Long userId) {
+        return placeJpaRepository.findAllByUserId(userId);
     }
+
     public List<Place> getAllByUserIdAndCountryName(Long userId, String countryName) {
         return placeJpaRepository.findAllByUserIdAndCountryName(userId, countryName);
     }
 
-    public Place savePlace (Place place) {
+    public Place savePlace(Place place) {
         return placeJpaRepository.save(place);
     }
+
     public void deletePlace(Long id) {
-        placeJpaRepository.deleteById(id);
+        Optional<Place> place = placeJpaRepository.findById(id);
+        if (place.isPresent()) {
+            placeJpaRepository.deleteById(id);
+        }
+
     }
 }
